+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-### Added
+### Added 
+- Add an `onClickOutside` event listener on modal, default is cancelation.
 - Added prop `iconPosition` for `buttonIcon`, can be *Left* (default) or *Right*
 
 ## [0.17.2] - 2019-09-20